requirements.txt
pandas==1.4.2
Pillow==9.4.0
torch==1.13.0
torchvision==0.14.0
tqdm==4.64.1
